+++ bobgui/modules/media/audio/bobguiaudio.h
@@ -19,3 +19,11 @@ BobguiAudioDeviceManager * bobgui_audio_device_manager_new (void);
 G_END_DECLS
 
 #endif
+void bobgui_audio_device_manager_init_devices(BobguiAudioDeviceManager *self, int input_channels, int output_channels);
+void bobgui_audio_device_manager_set_sample_rate(BobguiAudioDeviceManager *self, double sample_rate);
+void bobgui_audio_device_manager_play(BobguiAudioDeviceManager *self);
+void bobgui_audio_device_manager_stop(BobguiAudioDeviceManager *self);
+BobguiAudioProcessor * bobgui_audio_processor_new(const char *name);
+void bobgui_audio_processor_set_parameter(BobguiAudioProcessor *self, int index, float value);
+float bobgui_audio_processor_get_parameter(BobguiAudioProcessor *self, int index);
+void bobgui_audio_processor_process_block(BobguiAudioProcessor *self, float** in_channels, float** out_channels, int num_samples);