+++ bobgui/modules/media/shader/bobguishader.c
@@ -13,3 +13,5 @@ static void bobgui_effect_shader_class_init (BobguiEffectShaderClass *klass) {
 }
 
 BobguiEffectShader * bobgui_effect_shader_new_from_source (const char *source) { return NULL; }
+void bobgui_effect_shader_set_uniform_float(BobguiEffectShader *self, const char* name, float value) {}
+void bobgui_effect_shader_set_uniform_vec3(BobguiEffectShader *self, const char* name, float x, float y, float z) {}