+++ bobgui/modules/media/shader/bobguishader.h
@@ -13,3 +13,5 @@ BobguiEffectShader * bobgui_effect_shader_new_from_source (const char *source);
 G_END_DECLS
 
 #endif
+void bobgui_effect_shader_set_uniform_float(BobguiEffectShader *self, const char* name, float value);
+void bobgui_effect_shader_set_uniform_vec3(BobguiEffectShader *self, const char* name, float x, float y, float z);